feat: :sparkles: implement authorization

+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 use App\Exceptions\NotFoundException;
 use App\Http\Resources\CategoryResource;
 use App\Http\Requests\Category\{StoreCategoryRequest, UpdateCategoryRequest};
@@ -11,12 +12,14 @@ class CategoryController extends Controller
 {
     public function index()
     {
+        Gate::authorize('list_categories', Product::class);
         $categories = Category::paginate();
         return CategoryResource::collection($categories);
     }
 
     public function store(StoreCategoryRequest $request)
     {
+        Gate::authorize('new_category', Product::class);
         $category = Category::create($request->validated());
         return CategoryResource::make($category);
     }
@@ -24,12 +27,14 @@ class CategoryController extends Controller
     public function show($categoryId)
     {
         $category = $this->category($categoryId);
+        Gate::authorize('show_category', $category);
         return CategoryResource::make($category);
     }
 
     public function update($categoryId, UpdateCategoryRequest $request)
     {
         $category = $this->category($categoryId);
+        Gate::authorize('update_category', $category);
         $category->update($request->validated());
         return CategoryResource::make($category);
     }
@@ -37,7 +42,7 @@ class CategoryController extends Controller
     public function destroy($categoryId)
     {
         $category = $this->category($categoryId);
-        // dd($category->products);
+        Gate::authorize('destroy_category', $category);
         $category->delete();
         return response()->noContent();
     }
